Extract common test seeds
