+++ packages/ui/test/_mocks/server/seeds/index.ts
@@ -1,0 +1,79 @@
+import { getReward } from '@/working-groups/model/getReward'
+import { WorkerWithDetails } from '@/working-groups/types'
+
+import { alice } from '../../keyring'
+
+export const MEMBER_ALICE = {
+  id: '0',
+  rootAccount: '5GNJqTPyNqANBkUVMN1LPPrxXnFouWXoe2wNSmmEoLctxiZY',
+  controllerAccount: '5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY',
+  handle: 'alice',
+  metadata: {
+    name: 'animi et',
+    about: 'Optio possimus...',
+  },
+  isVerified: true,
+  isFoundingMember: true,
+  inviteCount: 5,
+  registeredAtBlockId: '1',
+}
+
+export const WORKER: WorkerWithDetails = {
+  membership: {
+    id: '1',
+    controllerAccount: alice.address,
+  },
+  group: {
+    name: 'forum',
+    id: '1',
+  },
+  isLeader: false,
+  reward: getReward(2, 'forum'),
+  earnedTotal: 2000,
+  stake: 2000,
+  status: '',
+  id: '1',
+  applicationId: '1',
+  openingId: '1',
+  roleAccount: '5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY',
+  stakeAccount: '5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY',
+  rewardAccount: '5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY',
+  hiredAtBlock: {
+    id: '100',
+    network: 'OLYMPIA',
+    number: 100,
+    timestamp: '2021-06-01T06:42:00.155Z',
+  },
+}
+
+export const OPENING = {
+  groupId: '0',
+  type: 'REGULAR',
+  status: 'open',
+  stakeAmount: 2000,
+  applications: null,
+  metadata: {
+    shortDescription: 'Distribution Worker',
+    description: '# Description',
+    hiringLimit: 1,
+    expectedEnding: '2022-03-09T10:18:04.155Z',
+    applicationDetails: 'Details... ?',
+    applicationFormQuestions: [
+      {
+        type: 'TEXT',
+        question: 'Question 1',
+      },
+      {
+        type: 'TEXT',
+        question: 'Question 2',
+      },
+      {
+        type: 'TEXT',
+        question: 'Question 3',
+      },
+    ],
+  },
+  unstakingPeriod: '14409',
+  rewardPerBlock: 200,
+  createdAtBlockId: '5',
+}
